All suggestions implemented and Changes made Accordingl To Latest Code Review
As suggested in today's Code Review Loom video, implemented almost all suggestions and Changes were made according to what was said. As all the Scripts including latest Key Controller and Score Controller Script were Reviewed, going to merge all branches. THANK YOU.
