+++ Assets/Prefab/Player.prefab
@@ -106,7 +106,6 @@ GameObject:
   - component: {fileID: 5683805777109378455}
   - component: {fileID: 3283899742976562841}
   - component: {fileID: 1427029619061578323}
-  - component: {fileID: 4140927669616488591}
   m_Layer: 3
   m_Name: Player
   m_TagString: Untagged
@@ -291,18 +290,3 @@ MonoBehaviour:
   animator: {fileID: 1534014322983105027}
   speed: 3
   jump: 26
---- !u!114 &4140927669616488591
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4060767588766888550}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f08a4f3b0eb202645a0f6d40ae4cb649, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  animator: {fileID: 1534014322983105027}
-  SceneID: 0
-  TimeDelay: 2
