+++ Assets/Scenes/Scene1.unity
@@ -168,7 +168,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 064ff2cb95452cd46abac22277169b16, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  animator: {fileID: 0}
   SceneID: 1
   TimeDelay: 2
 --- !u!61 &77412439
@@ -1326,6 +1325,26 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 569132788}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &728551221 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4060767588766888550, guid: 9430f5a651ff48645b75128227a87775,
+    type: 3}
+  m_PrefabInstance: {fileID: 653429294740093638}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &728551229
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 728551221}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 064ff2cb95452cd46abac22277169b16, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  SceneID: 0
+  TimeDelay: 2
 --- !u!1001 &851104266
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1438,7 +1457,7 @@ Transform:
   m_GameObject: {fileID: 1341735208}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 279.9, y: 125.9, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -2921,12 +2940,12 @@ PrefabInstance:
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: -5.9
+      value: 274
       objectReference: {fileID: 0}
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0.1
+      value: 126
       objectReference: {fileID: 0}
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
@@ -2968,10 +2987,15 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4140927669616488591, guid: 9430f5a651ff48645b75128227a87775, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 4060767588766888550, guid: 9430f5a651ff48645b75128227a87775,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 728551229}
   m_SourcePrefab: {fileID: 100100000, guid: 9430f5a651ff48645b75128227a87775, type: 3}
 --- !u!1001 &8762915958078302959
 PrefabInstance:
