+++ Assets/Scenes/Scene2.unity
@@ -168,7 +168,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 064ff2cb95452cd46abac22277169b16, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  animator: {fileID: 0}
   SceneID: 1
   TimeDelay: 2
 --- !u!61 &77412439
@@ -216,12 +215,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!95 &122824692 stripped
-Animator:
-  m_CorrespondingSourceObject: {fileID: 1534014322983105027, guid: 9430f5a651ff48645b75128227a87775,
-    type: 3}
-  m_PrefabInstance: {fileID: 653429294740093638}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &257915546
 GameObject:
   m_ObjectHideFlags: 0
@@ -1116,7 +1109,7 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 653429294740093638}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1305740807
+--- !u!114 &1305740808
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -1125,10 +1118,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 1305740801}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f08a4f3b0eb202645a0f6d40ae4cb649, type: 3}
+  m_Script: {fileID: 11500000, guid: 064ff2cb95452cd46abac22277169b16, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  animator: {fileID: 122824692}
   SceneID: 1
   TimeDelay: 2
 --- !u!1 &1341735208
@@ -1169,7 +1161,7 @@ Transform:
   m_GameObject: {fileID: 1341735208}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 55.9, y: 27.9, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -2294,7 +2286,7 @@ Transform:
   m_GameObject: {fileID: 2142229612}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 17.3, y: -8.3, z: 4.0613623}
+  m_LocalPosition: {x: 73.1, y: 17.8, z: 4.0613623}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -2313,15 +2305,20 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
+    - target: {fileID: 5366304130389893741, guid: 9430f5a651ff48645b75128227a87775,
+        type: 3}
+      propertyPath: m_Enabled
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: -5.9
+      value: 50
       objectReference: {fileID: 0}
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0.1
+      value: 28
       objectReference: {fileID: 0}
     - target: {fileID: 7484290047704761094, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
@@ -2363,14 +2360,15 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4140927669616488591, guid: 9430f5a651ff48645b75128227a87775, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 4060767588766888550, guid: 9430f5a651ff48645b75128227a87775,
         type: 3}
       insertIndex: -1
-      addedObject: {fileID: 1305740807}
+      addedObject: {fileID: 1305740808}
   m_SourcePrefab: {fileID: 100100000, guid: 9430f5a651ff48645b75128227a87775, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
