--- Assets/Scripts/DeathRestartController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f08a4f3b0eb202645a0f6d40ae4cb649
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
